+++ components/src/dynamo/vllm/multimodal_handlers/__init__.py
@@ -1,13 +1,13 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from dynamo.vllm.multimodal_handlers.encode_worker_handler import EncodeWorkerHandler
+from dynamo.vllm.multimodal_handlers.encode_worker_handler import (
+    EncodeWorkerHandler,
+    VLLMEncodeWorkerHandler,
+)
 from dynamo.vllm.multimodal_handlers.processor_handler import (
     ECProcessorHandler,
     ProcessorHandler,
-)
-from dynamo.vllm.multimodal_handlers.vllm_native_encoder_handler import (
-    VLLMNativeEncoderWorkerHandler,
 )
 from dynamo.vllm.multimodal_handlers.worker_handler import (
     MultimodalDecodeWorkerHandler,
@@ -16,9 +16,9 @@ from dynamo.vllm.multimodal_handlers.worker_handler import (
 
 __all__ = [
     "EncodeWorkerHandler",
+    "VLLMEncodeWorkerHandler",
     "ProcessorHandler",
     "MultimodalPDWorkerHandler",
     "MultimodalDecodeWorkerHandler",
-    "VLLMNativeEncoderWorkerHandler",
     "ECProcessorHandler",
 ]
