+++ examples/backends/vllm/launch/agg_multimodal_ec_connector.sh
@@ -105,7 +105,7 @@ python -m dynamo.vllm \
     --mm-prompt-template "$PROMPT_TEMPLATE" &
 
 # Start vLLM-native encoder worker (ECConnector producer)
-echo "Starting vLLM-native encoder worker (ECConnector producer) on GPU 0..."
+echo "Starting vLLM-native encoder worker (ECConnector producer) on GPU 2..."
 CUDA_VISIBLE_DEVICES=2 python -m dynamo.vllm \
     --vllm-native-encoder-worker \
     --enable-multimodal \
@@ -118,7 +118,7 @@ CUDA_VISIBLE_DEVICES=2 python -m dynamo.vllm \
     --no-enable-prefix-caching &
 
 # Start aggregated multimodal worker (ECConnector consumer, P+D combined)
-echo "Starting aggregated multimodal worker (ECConnector consumer) on GPU 1..."
+echo "Starting aggregated multimodal worker (ECConnector consumer) on GPU 3..."
 CUDA_VISIBLE_DEVICES=3 python -m dynamo.vllm \
     --multimodal-worker \
     --enable-multimodal \
